+++ CHANGELOG.md
@@ -43,6 +43,7 @@
     used for the creation of the endpoint
     ([issue #1251](https://github.com/PowerShell/SqlServerDsc/issues/1251)).
     [Maxime Daniou (@mdaniou)](https://github.com/mdaniou)
+- Added SqlAgentOperator Resource
 
 ## 12.2.0.0
 
@@ -113,7 +114,6 @@
   - Fix unit tests that didn't mock some of the calls. It no longer fail
     when a SQL Server installation is not present on the node running the
     unit test ([issue #983](https://github.com/PowerShell/SqlServerDsc/issues/983)).
-- Added SqlAgentOperator Resource
 
 ## 12.0.0.0